Refactor sample application
The Blazor Server App doesn't got the naming pattern.
