+++ sample/BlazorServer/BlazorServer.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>       
     </PropertyGroup>
 
     <ItemGroup>
