+++ sample/BlazorServer/Pages/FetchData.razor
@@ -1,6 +1,6 @@
 ﻿@page "/fetchdata"
 
-@using BlazorServerApp.Data
+@using BlazorServer.Data
 @inject WeatherForecastService ForecastService
 
 <h1>Weather forecast</h1>
