+++ sample/BlazorServer/Program.cs
@@ -9,7 +9,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace BlazorServerApp
+namespace BlazorServer
 {
     public class Program
     {
