+++ sample/BlazorServer/Shared/SurveyPrompt.razor
