+++ sample/BlazorServer/_Imports.razor
@@ -5,6 +5,6 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
-@using BlazorServerApp
-@using BlazorServerApp.Shared
+@using BlazorServer
+@using BlazorServer.Shared
 @using Wangkanai.Detection
