+++ sample/BlazorServer/appsettings.Development.json
