+++ sample/BlazorServer/wwwroot/css/open-iconic/FONT-LICENSE
