+++ sample/BlazorServer/wwwroot/css/open-iconic/ICON-LICENSE
