+++ sample/BlazorServer/wwwroot/css/open-iconic/README.md
