+++ sample/BlazorServer/wwwroot/css/open-iconic/font/fonts/open-iconic.otf
