+++ sample/BlazorServer/wwwroot/favicon.ico
